Fix error with resize label

+++ models/head/ct_head.py
@@ -93,7 +93,7 @@ class CT_Head(nn.Module):
         img_size = img_meta['img_size'][0]
 
         label_num = np.max(label) + 1
-        label = cv2.resize(label, (img_size[1], img_size[0]), interpolation=cv2.INTER_NEAREST)
+        label = cv2.resize(label, (int(np.float32(img_size[1])), int(np.float32(img_size[0]))), interpolation=cv2.INTER_NEAREST)
 
         if not self.training and cfg.report_speed:
             torch.cuda.synchronize()
